extract integration test to new dedicated class.
